requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
black==23.7.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
coverage==7.3.0
cucumber-tag-expressions==4.1.0
exceptiongroup==1.1.3
execnet==2.0.2
flake8==6.1.0
Flask==2.3.3
genson==1.2.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpath-ng==1.6.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
parse==1.19.1
parse-type==0.6.2
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.2.0
ply==3.11
py==1.11.0
pycodestyle==2.11.0
pyee==9.0.4
pyflakes==3.1.0
pytest==7.4.0
pytest-base-url==2.0.0
pytest-bdd==6.1.1
pytest-cov==4.1.0
pytest-html==4.0.2
pytest-metadata==3.0.0
pytest-xdist==3.3.1
python-slugify==8.0.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
six==1.16.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
zipp==3.16.2
